Optional padding lines on both sides
The main container often has a padding (e. g. 15px) on both sides, so that there is a gap between the container and the inner elements.
I added the possibility to show these padding, without messing with the columns (position: absolute).
You can add the properties paddingleft and paddingright to activate them.